+++ src/main/java/racingcar/domain/CarName.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 public class CarName {
     private final String name;
