+++ src/test/java/racingcar/domain/CarPositionTest.java
@@ -1,8 +1,9 @@
-package racingcar;
+package racingcar.domain;
 
 import static org.assertj.core.api.Assertions.*;
 
 import org.junit.jupiter.api.Test;
+import racingcar.domain.CarPosition;
 
 class CarPositionTest {
